% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{benchmark.data.scRNAseq}
\alias{benchmark.data.scRNAseq}
\title{Single-Cell RNA Sequencing Datasets}
\description{
This package contains nine benchmark scRNA-seq datasets. Seven are
sequenced from human samples, and two from mouse samples. The datasets
range from small (hundreds of cells) to large (hundreds of thousands
of cells). The datasets are stored in \code{SingleCellExperiment}
objects, and are available in \code{ExperimentHub}.
}
\examples{
library(ExperimentHub)
hub <- ExperimentHub()
x <- query(hub, c("benchmark.data.scRNAseq", "cellmix"))
x
\dontrun{
## download resource
cellmix = x[[1]]
}}
