% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hca_bm}
\alias{hca_bm}
\title{hca-bm Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://www.sciencedirect.com/science/article/pii/S0301472X18308051}
}
\usage{
data(hca_bm)
}
\description{
The hca-bm dataset was generated by Human Cell Atlas using
bone marrow cells from eight donors.
}
\details{
Species: Human

Cells/Tissue: Bone marrow of 8 donors

Platform: 10x (droplet-based)

Number of cells: 351,107

Number of genes: 9,651
}
\keyword{datasets}
