\docType{data}
\name{hca_cb}
\alias{hca_cb}
\title{hca-cb Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://data.humancellatlas.org/explore/projects/cc95ff89-2e68-4a08-a234-480eca21ce79}
}
\usage{
library(ExperimentHub)
hub <- ExperimentHub()
x <- query(hub, c("benchmark.data.scRNAseq", "pbmc"))
## download resource
pbmc = x[[1]]
}
\description{
The hca-cb dataset was generated by Human Cell Atlas using
cord blood cells
}
\details{
Species: Human

Cells/Tissue: Cord blood of 8 donors

Platform: 10x (droplet-based)

Number of cells: 352,406

Number of genes: 7,234
}
\keyword{datasets}
