% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{scRNAseq.datasets}
\alias{scRNAseq.datasets}
\alias{scRNA.seq.datasets}
\title{Single-Cell RNA Sequencing Datasets}
\description{
This package contains nine scRNA-seq datasets. Seven are
sequenced from human samples, and two from mouse samples.
The datasets range from small (hundreds of cells) to large
(hundreds of thousands of cells).
}
