% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zeisel}
\alias{zeisel}
\title{Zeisel Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://science.sciencemag.org/content/347/6226/1138}
}
\usage{
data(zeisel)
}
\description{
The Zeisel dataset was generated to classify cells in
the mouse somatosensory cortex and hippocampal CA1 region.
The authors found 47 molecularly distinct subclasses,
comprising all known major cell types in the cortex.
}
\details{
Species: Mouse

Cells/Tissue: Cortex (S1) and hippo- campus (CA1) tissue

Platform: STRT-Seq (microfluidics)

Number of cells: 3,005

Number of genes: 8,366
}
\keyword{datasets}
