\docType{data}
\name{hca_bm}
\alias{hca_bm}
\title{hca-bm Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://data.humancellatlas.org/explore/projects/cc95ff89-2e68-4a08-a234-480eca21ce79}
}
\usage{
library(ExperimentHub)
hub <- ExperimentHub()
x <- query(hub, c("benchmark.data.scRNAseq", "pbmc"))
## download resource
pbmc = x[[1]]
}
\description{
The hca-bm dataset was generated by Human Cell Atlas using
bone marrow cells from eight donors.
}
\details{
Species: Human

Cells/Tissue: Bone marrow of 8 donors

Platform: 10x (droplet-based)

Number of cells: 351,107

Number of genes: 9,651
}
\keyword{datasets}
