% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hca_cb}
\alias{hca_cb}
\title{hca-cb Dataset}
\format{
Large SingleCellExperiment
}
\source{
Human Cell Atlas
}
\usage{
data(hca_cb)
}
\description{
The hca-cb dataset was generated by Human Cell Atlas using
cord blood cells
}
\details{
Species: Human

Cells/Tissue: Cord blood of 8 donors

Platform: 10x (droplet-based)

Number of cells: 352,406

Number of genes: 7,234
}
\keyword{datasets}
