% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kolodziejczyk}
\alias{kolodziejczyk}
\title{Kolodziejczyk Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/26431182}
}
\usage{
data(kolodziejczyk)
}
\description{
The Kolodziejczyk dataset was generated using mouse Embryonic
Stem Cells (mESCs) cultured in three different conditions:
serum, 2i, and the alternative ground state a2i
}
\details{
Species: Mouse

Cells/Tissue: mESCs grown in 3 conditions

Platform: Fluidigm C1 (microfluidics)

Number of cells: 623

Number of genes: 21,453
}
\keyword{datasets}
