% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc}
\alias{pbmc}
\title{PBMC Dataset}
\format{
Large SingleCellExperiment
}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/1.1.0/fresh_68k_pbmc_donor_a}
}
\usage{
data(pbmc)
}
\description{
The PBMC dataset was generated by 10x Genomics using the
Cell Ranger 1.1.0 pipeline. The official name of the dataset
is "Fresh 68k PBMCs (Donor A)".
}
\details{
Species: Human

Cells/Tissue: Fresh cells, 1 donor

Platform: 10x (droplet-based)

Number of cells: 68,224

Number of genes: 3978
}
\keyword{datasets}
